Added description of GERBIL to entity_linking.md
Justification given for why such a benchmarking platform is necessary. Links
to papers on evaluating EL systems have been provided and a high level
explanation of GERBIL's features is given.

Reference has been made to NIF with links given to some documentation.

There is much more that could be said on all of the above, but this
initial commit at the very least helps to get the information into the
repository.